refactor(rollup): relocate config files into rollup/ directory
Move the five rollup configs (all/content/background/migration/visualizer)
out of the repository root and into rollup/, alongside the existing
shared/ and plugins/ subdirectories. Update internal imports, npm
build scripts, knip entry list, and release-please exclude-paths.

Also drop the stale `rollup.preloader.config.mjs` reference in
release-please-config.json (the preloader is a sub-config inside
rollup/all.config.mjs and never had its own file).

Verified byte-identical across all four production bundles after the
move (sha256 matches the pre-relocation baseline for content,
background, migration-executor, and preloader).
