+++ knip.json
@@ -2,11 +2,11 @@
   "$schema": "./node_modules/knip/schema.json",
   "ignoreExportsUsedInFile": true,
   "entry": [
-    "rollup.all.config.mjs",
-    "rollup.content.config.mjs",
-    "rollup.background.config.mjs",
-    "rollup.migration.config.mjs",
-    "rollup.visualizer.config.mjs",
+    "rollup/all.config.mjs",
+    "rollup/content.config.mjs",
+    "rollup/background.config.mjs",
+    "rollup/migration.config.mjs",
+    "rollup/visualizer.config.mjs",
     "pages/onboarding/*.js",
     "pages/options/*.js",
     "pages/popup/*.js",
