+++ package.json
@@ -23,13 +23,13 @@
     "format": "prettier --write \"**/*.{js,json,md,html}\"",
     "format:check": "prettier --check \"**/*.{js,json,md,html}\"",
     "postinstall": "node scripts/postinstall.js && npm run build",
-    "build": "rollup -c rollup.all.config.mjs",
-    "build:watch": "rollup -c rollup.all.config.mjs -w",
-    "build:prod": "NODE_ENV=production rollup -c rollup.all.config.mjs",
-    "build:prod:analyze": "ANALYZE_BUNDLE=true NODE_ENV=production rollup -c rollup.all.config.mjs",
-    "build:content": "rollup -c rollup.content.config.mjs",
-    "build:content:watch": "rollup -c rollup.content.config.mjs -w",
-    "build:content:prod": "NODE_ENV=production rollup -c rollup.content.config.mjs",
+    "build": "rollup -c rollup/all.config.mjs",
+    "build:watch": "rollup -c rollup/all.config.mjs -w",
+    "build:prod": "NODE_ENV=production rollup -c rollup/all.config.mjs",
+    "build:prod:analyze": "ANALYZE_BUNDLE=true NODE_ENV=production rollup -c rollup/all.config.mjs",
+    "build:content": "rollup -c rollup/content.config.mjs",
+    "build:content:watch": "rollup -c rollup/content.config.mjs -w",
+    "build:content:prod": "NODE_ENV=production rollup -c rollup/content.config.mjs",
     "build:all": "npm run build",
     "build:all:prod": "NODE_ENV=production npm run build:prod",
     "dev": "npm run build:watch",
