+++ rollup/content.config.mjs
@@ -1,12 +1,12 @@
 import resolve from '@rollup/plugin-node-resolve';
 import commonjs from '@rollup/plugin-commonjs';
 import replace from '@rollup/plugin-replace';
-import { createVisualizerPlugin } from './rollup.visualizer.config.mjs';
-import { isDev } from './rollup/shared/env.mjs';
-import { createTerserPlugin } from './rollup/shared/terser.mjs';
-import { createOnWarn } from './rollup/shared/onwarn.mjs';
-import { stripTestConfig } from './rollup/plugins/stripTestConfig.mjs';
-import { assertTestFixtureDce } from './rollup/plugins/assertTestFixtureDce.mjs';
+import { createVisualizerPlugin } from './visualizer.config.mjs';
+import { isDev } from './shared/env.mjs';
+import { createTerserPlugin } from './shared/terser.mjs';
+import { createOnWarn } from './shared/onwarn.mjs';
+import { stripTestConfig } from './plugins/stripTestConfig.mjs';
+import { assertTestFixtureDce } from './plugins/assertTestFixtureDce.mjs';
 
 export default {
   input: 'scripts/content/index.js',
